show debug tile, fix blow position

+++ cam.py
@@ -14,10 +14,10 @@ class Cam:
         self.pos_y = 0
 
     def gridToScreen(self, x, y):
-        return x * TILE_W - self.pos_x, y * TILE_H - self.pos_y
+        return int(x * TILE_W - self.pos_x), int(y * TILE_H - self.pos_y)
 
     def screenToGrid(self,x, y):
-        return (x + self.pos_x) // TILE_W, (y + self.pos_y) // TILE_H
+        return int((x + self.pos_x) // TILE_W), int((y + self.pos_y) // TILE_H)
 
     def scroll(self, bounds, scroll_directions):
         self.pos_x += scroll_directions[0] * SCROLL_SPEED
